move test to /test/. Add build tag.
